--- aemedge/blocks/blogs-by-author/blogs-by-author.css
@@ -1,160 +1,0 @@
-
-.pagination-container {
-    align-items: center;
-    display: flex;
-    justify-content: space-between;
-}
-
-@media screen and (width >=768px) {
-.pagination-container {
-    align-items: center;
-    display: flex;
-    justify-content: space-between;
-    margin: 0 auto 134px;
-    max-width: 1166px;
-}
-}
-
-.pagination-container button.paginate-btn{
-    height: 50px;
-    font-size: 16px;
-    line-height: 52px;
-    width: 168px;
-    text-align: center;
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-.pagination-info-text {
-    font-size: 16px;
-}
-
-.pagination-container .paginate-btn.inactive {
-    background: white;
-    border: 1px solid #0098d1;
-    color: #0078ad;
-    cursor: default;
-}
-
-.pagination-container .paginate-btn.active {
-    background: #0098d1;
-    color: white;
-    cursor: pointer;
-}
-
-.cards-container {
-    margin: 16px auto;
-}
-
-@media screen and (width >=768px) {
-.cards-container {
-    margin: 0 auto;
-    max-width: 1166px;
-}
-}
-
-.card-container {
-    margin: 16px auto;
-    width: 370px;
-}
-
-@media screen and (width >=768px) {
-    .card-container { 
-        display: flex;
-        margin: 32px 0;
-        width: 100%;
-    }
-}
-
- .image-container {
-    border-radius: 4px;
-    margin-right: 46px;
-    box-sizing: border-box;
-}
-
-.card-container .image-container .image-link {
-    cursor: pointer;
-    height: 100%;
-    max-width: 367px;
-}
-
-.card-image {
-    border-radius: 4px;
-    max-height: 207px;
-    max-width: 367px;
-    vertical-align: middle;
-}
-
-.text-container {
-    margin-top: 16px;
-    margin-bottom: 16px;
-}
-
-@media screen and (width >=768px) { 
-    .text-container {
-        max-height: 70px;
-    }
-}
-
-a.title-link{
-    color: #000;
-    font-size: 22px;
-    font-weight: 700;
-    line-height: 27px;
-    margin: 0 auto;
-    text-align: center;
-    word-wrap: break-word;
-}
-
-@media screen and (width >=768px) {
-.text-container {
-    display: inline-block;
-    max-height: 160px;
-}
-
-a.title-link{
-    color: #000;
-    font-size: 22px;
-    font-weight: 700;
-    line-height: 27px;
-    margin-bottom: 8px;  
-}
-}
-
-.description-container { 
-    display: none;
-}
-
-@media screen and (width >=768px) {
-.description-container {
-    color: #1d1d1d;
-    font-size: 16px;
-    max-width: 621px;
-    display: block;
-}
-}
-
-.date-container {
-    display: inline-flex;
-    margin-top: 15px;
-    max-height: 21px;
-    max-width: 368px;
-}
-
-.author-container {
-    font-size: 14px;
-    line-height: 21px;
-    color: #171725;
-}
-
-.publication-date {
-    margin-left: 10px;
-}
-
-hr.card-separator {
-    border: 0;
-    border-top: 1px solid #454550;
-    max-width: 1166px;
-    margin: 0;
-}
